chrfi : small bug fix in load method (doulbe for size instead of int)
git-svn-id: http://rsg.pml.ac.uk/intranet/svnroots/arsf-internal/trunk/tools/lag@611 e4e130c3-e29c-4adf-81f0-605662ec9256

+++ quadtree.cpp
@@ -329,7 +329,7 @@ void quadtree::load(lidarpointloader *loader, int nth)
 
 
 // this method loads points from a flightline that fall within an area of intrest
-void quadtree::load(lidarpointloader *loader, int nth, double *Xs, double *Ys, double size)
+void quadtree::load(lidarpointloader *loader, int nth, double *Xs, double *Ys, int size)
 {
    /*double a1,a2,a3;
